refactor: move awaiter to separate file
